Page Update [68] | Update Context Api | Part Update [36]
Within the scope of the UI/UX Design Overhaul (Phase 8), Page Update [68], Update Context API, and Part Update [36] have been implemented. This update includes enhancements to the login page, as well as other theme components, and updates to the Context API. Many sections have been connected to the Context API.

+++ mogart-network/src/MogartBase/ThemeParts/MainPart/Navbar/Navbar.tsx
@@ -10,6 +10,7 @@ import Notification, { MessageType } from '../../Notification/Notification';
 
 export default function Navbar() {
   const { isLoggedIn, data } = useData();
+  const [rememberMe, setRememberMe] = useState(false);
   const [notification, setNotification] = useState({ show: false, type: MessageType.Info, message: '' });
   const navigate = useNavigate();
   
@@ -40,13 +41,19 @@ export default function Navbar() {
   
     try {
       const savedUserAuthID = localStorage.getItem('userAuthID');
+      const rememberemail = localStorage.getItem('rememberuserEmail');
       const response = await logout({ userid: savedUserAuthID, email: data.Email, walletadress: data.walletAddress });
   
       if (response.success === false) {
         showNotification(MessageType.Error, response.message);
       } else {
         showNotification(MessageType.Success, "Logout successful");
-        localStorage.clear();
+        if(rememberemail)
+        {
+          localStorage.clear();
+          localStorage.setItem('rememberuserEmail', rememberemail);
+        }else { localStorage.clear();}
+        
         navigate("/login");
       }
     } catch (error) {
